feat: complete vote detail page

+++ back-end/src/controller/vote-controller.js
@@ -9,17 +9,18 @@ class VoteController {
     ctx.body = result
   }
 
-  async getVoteByName(ctx, next) {
-    const user_id = ctx.user.id
-    const [result] = await voteService.getVoteByName(user_id)
+  async getVote(ctx, next) {
+    const [result] = await voteService.getVote()
     ctx.body = result
   }
 
   async getVoteByVoteId(ctx, next) {
     const vote_id = ctx.params.id
-    const [result1] = await voteService.getVoteInfo(vote_id)
-    const [result2] = await voteService.getVoteOptions(vote_id)
-    ctx.body = result1.concat(result2)
+    const [info] = await voteService.getVoteInfo(vote_id)
+    const [voted] = await voteService.getVoteStatus(vote_id)
+    const [options] = await voteService.getVoteOptions(vote_id)
+    const [users] = await voteService.getVotedUsers(vote_id)
+    ctx.body = {info: info[0], voted, options, users}
   }
 
   async voteOption(ctx, next) {
@@ -30,18 +31,22 @@ class VoteController {
     const [voteInfo] = await voteService.getVoteInfo(vote_id)
     if (voteInfo.length) { // 查看是否有这个投票
       const multiple = voteInfo[0].multiple
-      const [voted] = await voteService.isVotedOption(user_id, vote_id, option_id)
       if (multiple === "1") {  // 是否是多选
+        const [voted] = await voteService.isVotedMultiple(user_id, vote_id, option_id)
         if (voted.length) {  // 是否已经选过
           result = await voteService.deleteOption(user_id, vote_id, option_id)
         } else {
           result = await voteService.addOption(user_id, vote_id, option_id)
         }
       } else {
+        const [voted] = await voteService.isVotedSingle(user_id, vote_id)
         if (voted.length) {
-          result = await voteService.deleteOption(user_id, vote_id, option_id)
+          if (voted.option_id === option_id) {
+            result = await voteService.deleteOption(user_id, vote_id, option_id)
+          } else {
+            result = await voteService.updateOption(user_id, vote_id, option_id)
+          }
         } else {
-          await voteService.deleteReverseOption(user_id, vote_id, option_id)
           result = await voteService.addOption(user_id, vote_id, option_id)
         }
       }
